Add files via upload
The core files of the structures sizing script. NOTE: to anyone who tries to run this code, it will take more setup for it to function as intended. There are file paths in the code that will have to be changed to a local location on your computer, or somewhere accessible to your computer. You will also require the Propulsion 1DOF model for their point designs in order to get the proper inputs for the script.